[dpwa] Fix the bug that existing tests are replaced just because of the name change.
In the original change, both test id, which is equivalent to the test steps, and test name are used to determine if a test should be deleted / added.
i.e. both of them need to be an exact match in order for a test to be
kept. But this causes a huge problem when new tests are generated
because all the existing tests with a different naming convention will
be updated to the new one.
Now, only use the test id for comparison, so the existing tests are
not impacted and the new tests will have the new naming convention.

This CL also fixed a small issue that test id is incorrectly generated when the code
of an action spans over 1 line. This is because `\n` is used to
breakdown the test steps.
Unit tests are also update to cover this issue.

Bug: b/307320179
Change-Id: Ia50d412657f17f737ab6df331a8ee233c83c5936
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4977543
Commit-Queue: Clifford Cheng <cliffordcheng@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215207}