[Extensions] Fix incognito enablement for dynamic scripts
UserScript objects need to be marked as enabled in incognito in
order to inject in incognito processes. This is properly done for
manifest-specified content scripts and newly-registered dynamic
content scripts, but is not properly set when loading dynamic content
scripts from storage or for the new userScripts API.

Move setting this bit to where we construct a UserScript from a
SerializedUserScript so that it applies to all these sites
appropriately.

The right fix for this is to really pull incognito enablement off
UserScripts entirely and just have it associated with the extension,
but that's a bit of a larger fix.

Add a regression test for dynamic scripts and incognito.

Bug: 1494155, 1495191
Change-Id: I70659777447f93c4cbadee1a219966222428887b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4968611
Reviewed-by: Kelvin Jiang <kelvinjiang@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215220}