Show placeholder value when previewing empty fields.
Initially suggested values were hidden/shown based on whether the
underlying value is empty or not. crrev.com/c/4769360 changed this so
that we check whether the underlying value is null or not, so that
Autofill could sometimes preview empty fields.

Consider a field <input type=text placeholder=Foo> and Autofill offers
the user to fill the value "Bar" into that field. Before the user
accepts this suggestion, Autofill displays the suggested value Bar in
the field. The new Undo Autofill feature enables the user to revert
the field's value to the empty value. In that case, the suggested
value is the empty string. When previewing this empty string, the
placeholder should be visible.

The solution is to let only the "visibility" of the placeholder
"element" depend on the nullity of the suggested value, but all other
attributes (value or style properties) will depend on its emptiness.

This CL:
-Fixes the mentioned problem by letting the placeholder visibility vis a
vis the suggested value depend on the nullity of the value and not its
emptiness.
- Removes unused functions that are supposed to be used for placeholder
visibility but aren't

Bug: 1441410
Change-Id: I55ce16af879e4565e68d9baa59accd844dbfaf32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4952696
Reviewed-by: Mason Freed <masonf@chromium.org>
Auto-Submit: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1215232}