Invoke root view onApplyWindowInsets in InsetObserver
This applies the view's default behavior[1] after we've gotten a chance to:
a) Modify the insets
b) Observe and stash them

This should complete the puzzle of insets application in InsetObserver.
Before https://crrev.com/c/4970757, we invoked the default
behavior, but didn't give get a chance to modify the insets in-flight.
As of https://crrev.com/c/4970757, we modify the insets but don't
properly apply them afterwards.
As of this CL, we should be both modifying and applying correctly.

[1] https://developer.android.com/reference/android/view/View.OnApplyWindowInsetsListener

Bug: 1495675
Change-Id: I9d3affc7952c72b9b7546c0423e9db571ff5322c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4977600
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215238}