Avoid comparing class name returned by View::GetClassName() by pointer
This CL fixes the callers of View::GetClassName() to use string_view to
compare two strings.

Previously, View::GetClassName() returned a pointer to a static constant
string, but in a past CL this was changed and now the pointer can point
to the internal buffer of std::string. The caller should not depend on
the implementation details anyways and we should use std::string,
std::string_view, or strcmp to compare class names.

Bug: b:306495485
Change-Id: I15d5c1af41761e337c818340c191d4fad5bd39c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4974762
Reviewed-by: Yuki Awano <yawano@google.com>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Jun Ishiguro <junis@google.com>
Cr-Commit-Position: refs/heads/main@{#1215300}