[autopip] Use Blink's activation lifespan directly instead of copying
Currently, we have a constant in AutoPictureInPictureTabHelper of 5
seconds to represent the length of a user gesture. While a user gesture
is generally 5 seconds, on address-sanitizer builds, it's changed to 60
seconds to accommodate how slow those binaries are. Since the
AutoPictureInPictureTabHelper was always using 5 seconds, some of its
tests are flaky on address-sanitizer builds. To fix this, this CL
changes the AutoPictureInPictureTabHelper to directly use the constant
from Blink instead of its own, and therefore gets the longer 60-second
timer on address-sanitizer builds.

Bug: 1485641
Change-Id: Ia83943b62d8efb156e278cdab0d3f7e3ed27d528
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4977751
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215337}