PlzDedicatedWorker: fix possible nullptr assignment.
As far as I took a look at the code, I suppose there can be nullptr
returned by `WorkerDevToolsManager::GetDevToolsHostFromToken`.
However, the callee looks using the returned value without ensuring
non-nullptr.  I suppose it can cause the crash in crbug.com/1265325.

As far as I checked where the token comes, it looks coming from
`DedicatedWorker::ContinueStart`.  At a glance, I misunderstood
that it can be newly created in `DevToolsAgent::WorkerThreadCreated`,
which happens only for Worklet, and not for DedicatedWorker.
Since `DedicatedWorker::Start` should record the `DedicatedWorkerHost`
`WorkerDevToolsManager::GetDevToolsHostFromToken` should be able to
find it.

The exceptional case is a `DedicatedWorkerHost` destruction.
If the destruction happens while `ChildTargetCreated` is inflight,
`WorkerDevToolsManager::GetDevToolsHostFromToken` may return nullptr.
I believe we can ignore it in that case because it is not possible
to continue.

Bug: 1265325
Change-Id: I12905d7a906bb0af5c4cbbe36f2a6b394e7d56b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4970619
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215355}