media/gpu/v4l2/test: Correct VP9 compressed header handling
The selection of compressed header support is done when the
Vp9Parser is instantiated. This was set to true always,
meaning that the Vp9Parser needed to parse the compressed
headers. Some V4L2 drivers can support compressed header
parsing. This support was being checked immediately afterwards
with the |supports_compressed_headers_| variable.

Moving the check to before the instantiation allows the
Vp9Parser to correctly specify whether the headers need
to be parsed.

Bug: b:307350016

Change-Id: I7f585eaf5f413780274431bd44a4cb9d95f4a252
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4978094
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215359}