[waffle]Fix omnibox being hidden in FRE on launch
The Desktop FRE is rendered in a Profile Picker with a default height
of 758px. We hide the fake omnibox in the UI for heights that are
less than 768px. We decided to change this value to 758px.

Screen recording: http://b/302108545#comment10

Bug: b:302108545
Change-Id: If1582ecd957c6e20b19946f1beff6340a86e725e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4974325
Auto-Submit: Jack Yammine <jyammine@google.com>
Commit-Queue: Nicolas Dossou-Gbété <dgn@chromium.org>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215428}