gpu: add EnablePerContextGLTextureCache for OzoneImageBacking.
There can be multiple of raster tasks that produce skia representation
over and over again, which triggers creation/destruction of glTexture
and the underlying EGLImage.

The idea behind the EGLImage is that it's reused/updated almost every
single frame. Reconstruction of that is not efficient. For example,
on LaCros, it takes about 200-400us to construct an EGLImage, which
makes the gpu thread more busy with the raster tasks (there can be
multiple of those as LaCros needs to rasterize to EGLImages).

A EnablePerContextGLTextureCache will make OzoneImageBacking store
TextureHolders on a per context basis.

The feature is disabled by default and will be enabled in a next
CL to ensure a smooth transition.

Test: OzoneImageBackingFactoryTest.UsesCacheForTextureHolders
Bug: 1479884
Change-Id: Id43a8aad7c8e28a10149e24910c869c286689b60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4965602
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1215432}