[gardener] Disable PrerenderBookmarkBarOnHoverNavigationTest.PrerenderNonHttps.
This test is consistently failing on bots, and occasionally when run
locally (about 10%). This is not new; I've gone back to at least 2023-10-10,
and it is failing there as well, so I suspect it was flaky from the point
where it was introduced. Disable it to keep the bots happy.

Change-Id: Id62b493c1c1fd31a1f08566113a81bc2c11abebd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4976154
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Owners-Override: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215438}