webview: Expose AwContentsOriginMatcher to Java
We want to make use of this class in Java land so that we can
have consistent matching rules.
Rather than re-implement this behaviour in Java, we are adding a
simple wrapper class to call into a native implementation.

The Java class doesn't keep onto any state.

The one unfortunate detail is that callers of this API will have
to call the destroy method to clear the native memory.

Bug: 1493963
Change-Id: Ib1735303350017074b5c2f6b1a288a19999f4ab8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4973035
Reviewed-by: Peter Conn <peconn@chromium.org>
Auto-Submit: Rupert Wiser <bewise@chromium.org>
Commit-Queue: Rupert Wiser <bewise@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215442}