[Sync][Lacros] Fix  TwoClientWebAppsBMOSyncTest
FixTwoClientWebAppsBMOSyncTest sync_integration_tests when kSyncChromeOSAppsToggleSharing is enabled. On Lacros, if kSyncChromeOSAppsToggleSharing is enabled, syncing of Apps is controlled via the OS, not the browser.
Activate the Apps toggle for both clients during the test.

Bug: 1330894 , b:307483579
Change-Id: Ic140714bd3b09d97cbb4ce46a6e88c6ae3726342
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4976230
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Brahim Chikhaoui <bchikhaoui@google.com>
Cr-Commit-Position: refs/heads/main@{#1215443}