[Sync][Lacros] Fix  TwoClientLacrosWebAppsSyncTest.SyncDisabledUnlessPrimary
Fix TwoClientLacrosWebAppsSyncTest.SyncDisabledUnlessPrimary Sync_integration_tests when kSyncChromeOSAppsToggleSharing is enabled. On Lacros, if kSyncChromeOSAppsToggleSharing is enabled, syncing of Apps is controlled via the OS, not the browser.
Activate the Apps toggle for both clients during the test and assert that only for the main profile's Apps toggle data types are included.

Bug: 1330894 , b:307483579
Change-Id: I4023e2560e4350b59d9c9024c96cb57ce8ab446b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4974993
Commit-Queue: Brahim Chikhaoui <bchikhaoui@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215454}