android: explicitly destroy the stencil framebuffer.
gl::GLContext holds offscreen surfaces now. However, they are not
destroyed until very end...until the context is actually destroyed.

It's OK in most of the cases, but it is not good in case of
the AwGLSurfaceExternalStencil that subclasses the AwGLSurface and that
creates stencil framebuffers that has to be destroyed on a correct
context.

Given that surface is also offscreen one (changed in
https://crrev.com/c/4942195), gl::GLContext holds it and it's not
guaranteed that particular context will be current. As a result, no
context/wrong context crash may happen.

To mitigate the issue, add a helper stencil framebuffer destruction
method that AwGLSurfaceExternalStencil implements, and which is called
when OutputSurfaceProviderWebView no longer needs the surface.

Bug: 1494632, 1479884
Change-Id: Ic8eecf7ee5d3a300e29bc21c07e6c26da66b8e00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4977718
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1215481}