[ios] Fix for leading icon animation on omnibox focus
Previously the leading icon was shifted with the omnibox text as part
of the focus animation, so in addition to the scale-in it was also
moving. This is not what UX wanted. This change modifies the animation
to only shift the "text field" rather than the whole "edit view" (which
includes the leading icon). This allows the scale-in effect to start
sooner, which gives it more time to complete and makes it look more
smooth. And it more closely matches the UX spec.

Mock from UX:
https://drive.google.com/file/d/1QkVV6hKNRAaqTq7tOUlswWRyB_AsquJZ/view?resourcekey=0-qS27kzTEdGAbJAKNmjC3iw

Video with changes:
http://screencast/cast/NTQ0MDMwMDA4NzUwODk5MnwwZGIyYjBlZi0xMg

Bug: 1482304
Change-Id: Idfbd72fa976f84f6b70ab410bd02fe89e3fe66a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4941012
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1215498}