media: Add "Save video frame as" support
This CL adds a context menu item "Save video frame as" to the <video>
element which downloads the current video frame as a PNG file. The
user experience is the same as "Save image as" for <image> and <canvas>.

This feature is still going through the launch process. So landed it
behind a disabled-by-default feature.

Manually tested on Windows and Mac and the video frame downloaded as
a PNG file in its original resolution. On Windows, it also works when
using hardware decoder (D3D11VideoDecoder). Also tested with protected
content and the option is disabled.

See screenshots at https://photos.app.goo.gl/zsi47KrdntYqLtTQ9

Bug: 1466685
Test: See above
Change-Id: Ia1acd243f79f50650989c90550ed649691577cfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4969014
Reviewed-by: Evan Liu <evliu@google.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215520}