Parse, but do not use, trigger_specs' trigger_data/event_report_windows
https://wicg.github.io/attribution-reporting-api/#parse-trigger-specs

We represent the specs as a map from trigger data values to indexes into
a list of specs. This representation reduces the amount of data in
memory while allowing it to be efficiently serialized in the Attribution
Reporting database using the same representation (which will be done in
a separate CL), and simultaneously allows duplicates and excessive
cardinality to be detected efficiently during parsing.

OBSOLETE_HISTOGRAM[Conversions.SourceRegistrationError6]=Replaced by Conversions.SourceRegistrationError7

Bug: 1495162
Change-Id: I2775876efdbbdd140f99b4525344b4852eb56320
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4973896
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215566}