Fixing a host config parsing bug on Windows
It turns out that Windows has some extra host config parsing logic
that rejects a config if it doesn't have a hash_secret. We just
need to skip the check if the owner is a Googler, just like we do
in the me2me host however instead of duplicating the same email
checking logic, I moved it into a helper and updated the other
//remoting/host callsite which checks the host_owner_email to see
if they are a Googler.

Bug: b:297857922
Change-Id: Ie104a1be299e67dcef46bd4aee85189d5e66cbbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4977281
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215567}