[DTC] Allow starting the inline flow from redirects
When coming from a redirect to one of the watched URLs, the response
headers will not be null. Therefore, offer the inline flow
(X-Device-Trust header) in the redirected request if the response
headers don't contain a Challenge (at which point the challenge
response should be generated instead).

Moved the funnel histogram validation from the browser tests to the
unit tests, as the embedded test server was retrying 3 times for some
tests (returning 307s) before starting to feed the requests through
the registered handler function.

Bug: b:307571482,1495859
Change-Id: Idef5d98077839c05bdfe06bc3aaef3b96afd8ef1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4977079
Reviewed-by: Hamda Mare <hmare@google.com>
Commit-Queue: Sébastien Lalancette <seblalancette@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1215580}