Extension port: only unregister the BFCached frames after navigation
In https://crrev.com/c/4914116, we call `port_->CloseChannel()` when
the FrameTracker observes a navigation that puts a page into BFCache.

This may cause problem if the navigation is a prerendering activation
because of two issues:
1. In the current implementation, when a navigation happens, the port
will unregister the new frame unless it's a same-doc navigation or
BFCache restoration. So the port created during prerendering will not
work after activation.
2. Previously we will close the extension port when a navigation that
puts a page into BFCache is committed in the WebContents. This assumes
the port only has registers the frames from the BFCached page. This is
not true with prerendering.

In this CL, we make the following changes to resolve the two issues:
1. Change the UnregisterFrame check from
`!IsServedFromBackForwardCache()` to `!IsPageActivation()` so
prerendered frame won't be unregistered.
2. When a navigation puts a page into BFCache, unregister all the frames
under the BFCached frame instead of closing the channel.

Bug: 1493966, 1472186
Change-Id: Icaaea036c1d7498f4c8da284e426d22c891a7de0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4964540
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215598}