Implement SyncFeatureStatusForMigrationsRecorder
SyncFeatureStatusForMigrationsRecorder is a helper class that tracks
the status of Sync-the-feature overall, as well as all individual
data types, and saves the results in prefs. This should be useful for
data migrations that (a) run early during browser startup, before Sync
gets initialized, and (b) must only run for users where Sync is
working correctly.

Note: The status here encompasses only whether that Sync-the-feature
(or the data type in question) has successfully become ACTIVE. In particular, this implies that it's not disabled due to either
configuration (user hasn't opted in, or has disabled a data type, etc)
or due to various error cases, e.g. enterprise policy, which may be
hard to detect otherwise during browser startup.

Bug: 1486420
Change-Id: I76cbaab91a49788985fc00175535e3a69921a751
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4946210
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1215608}