Deflake testScreenshotUserCounts
I started deflaking this in crrev.com/c/4963549 but
I'm still seeing some flakes in the wild.
"MobileClipboardChanged" is showing up as the result in some
tests. When "MobileClipboardChanged" is searched for, it is
added to user actions in ClipboardImpl.java. So I think the
issue is this test assumes that "Tab.Screenshot" is the first
action in the user action list. I think it should be acceptable
to just check the list contains "Tab.Screenshot" rather than
the first item being "Tab.Screenshot".

Bug: 1495179
Change-Id: Ia2c1bed45a9042a490a04865efa1cd4872fecc20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4977744
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1215611}